Write u-test for afro esc get methods. Refactor conversion logic.

+++ drone/src/lps25h.cpp
@@ -11,7 +11,7 @@ double Lps25h::get_pressure()
     uint8_t* buf = _get_buffer(LPS25H_PRESS_OUT_XL);
 
     int32_t raw_pres = ((buf[LPS25H_BUF_PRESS_OUT_H] << 16) | (buf[LPS25H_BUF_PRESS_OUT_L] << 8)) | buf[LPS25H_BUF_PRESS_OUT_XL];
-    double pres = double(raw_pres) * double(LPS25H_PRES_SCALE) / double(LPS25H_PRES_RESOLUTION);
+    double pres = double(raw_pres) * LPS25H_PRES_SCALE / LPS25H_PRES_RESOLUTION;
 
     return pres;
 }
@@ -22,7 +22,7 @@ double Lps25h::get_temperature()
     uint8_t* buf = _get_buffer(LPS25H_TEMP_OUT_P_L);
 
     int16_t raw_temp = (buf[LPS25H_BUF_TEMP_OUT_P_H] << 8) | buf[LPS25H_BUF_TEMP_OUT_P_L];
-    double temp = double(raw_temp) / double(LPS25H_TEMP_RESOLUTION) + double(LPS25H_TEMP_OFFSET);
+    double temp = double(raw_temp) / LPS25H_TEMP_RESOLUTION + LPS25H_TEMP_OFFSET;
 
     return temp;
 }
